puedes transformar la galeria absurda en un carrousel 3d manteniendo la
