+++ src/app/globals.css
@@ -95,6 +95,25 @@
   .rotate-y-180 {
     transform: rotateY(180deg);
   }
+  .carousel-container {
+      perspective: 1000px;
+  }
+  .carousel {
+      position: relative;
+      width: 220px;
+      height: 300px;
+      margin: 0 auto;
+      transform-style: preserve-3d;
+      transition: transform 1s;
+  }
+
+  .carousel-card {
+      position: absolute;
+      width: 100%;
+      height: 100%;
+      backface-visibility: hidden;
+      -webkit-backface-visibility: hidden;
+  }
 }
 
 @keyframes shimmer {
